temporary fix due to redefinition of gamma
this fix will make the table output consistent with the new draft of the paper (which defines gamma*Z as the level of the rate and not as the spread between the z.b. rate and the t-bill rate). This is temporary the permanent fix will make the code's definition of gamma consistent with the paper.